% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootBTdecayLasso.R
\name{boot.BTdecayLasso}
\alias{boot.BTdecayLasso}
\title{Compute the standard deviation of Bradley-Terry decay Lasso model by bootstrapping}
\usage{
boot.BTdecayLasso(dataframe, ability, lambda, boot = 100, weight = NULL,
  decay.rate = 0, fixed = 1, thersh = 1e-05, max = 100, iter = 100)
}
\arguments{
\item{dataframe}{Matrix with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).
It can be generated using function BTdataframe.}

\item{ability}{A column vector of teams ability, the last row is the home parameter.
The row number is consistent with the team's index shown in dataframe.
It can be generated using function BTdataframe.}

\item{lambda}{The amount of Lasso penalty induced, only a single scalar is accepted in bootstrapping.}

\item{boot}{Amount of simulations.}

\item{weight}{Weight for Lasso penalty on different abilities.}

\item{decay.rate}{The exponential decay rate. Usually ranging from (0, 0.1), A larger decay rate weights more
importance to most recent matches and the estimated parameters reflect more on recent behaviour.}

\item{fixed}{A teams index whose ability will be fixed as 0 (usually the team loss most which can be
generated using function BTdataframe).}

\item{thersh}{Thershold for convergency}

\item{max}{Maximum weight for w_{ij} (weight used for Adaptive Lasso)}

\item{iter}{Number of iterations used in L-BFGS-B algorithm.}
}
\value{
A list contain Lasso and Hybrid Lasso's bootstrapping's mean and standard deviation.
\item{Lasso}{Lasso bootstrapping's result. A three column matrix where first column is the original dataset's estimation, the second column is bootstrapping mean and the last column is the
bootstrapping standard deviation}
\item{HYBRID.Lasso}{HYBRID Lasso bootstrapping's result. A three column matrix where the first column is the original dataset's estimation, the second column is bootstrapping mean and the last column is the
bootstrapping standard deviation}
}
\description{
Compute the standard deviation of Bradley-Terry decay Lasso model by bootstrapping
}
\details{
100 times of simulation will be done by default, user can adjust the numbers of simulation by input of boot. However, bootstrapping process
is time consuming and in order to produce stable result, 1000 times of simulations is enough.
}
\examples{
##Initialize the dataframe and ability
NFL <- BTdataframe(NFL2010)

\dontrun{
##Run Lasso estimate for whole Lasso path
z <- BTdecayLasso(NFL$dataframe, NFL$ability, fixed = NFL$worstTeam)

##Model selection using AIC with Lasso's likelihood
z1 <- BTdecayLassoC(NFL$dataframe, NFL$ability, model = z, 
                    criteria = "AIC", type = "LASSO", fixed = NFL$worstTeam)

##Bootstrapping for model with lowest AIC score for 100 times.
z2 <- boot.BTdecayLasso(NFL$dataframe, NFL$ability, lambda = z1$Optimal.lambda, 
                        boot = 100, fixed = NFL$worstTeam)
}

}
\references{
Masarotto, G. and Varin, C.(2012) The Ranking Lasso and its Application to Sport Tournaments. 
*The Annals of Applied Statistics* **6** 1949--1970.

Zou, H. (2006) The adaptive lasso and its oracle properties. 
*J.Amer.Statist.Assoc* **101** 1418--1429.
}
\seealso{
\code{\link{BTdataframe}} for dataframe initialization
}
