% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTdataframe.R
\name{BTdataframe}
\alias{BTdataframe}
\title{Dataframe initialization}
\usage{
BTdataframe(dataframe)
}
\arguments{
\item{dataframe}{Raw dataframe input, an example data is given as NFL2010 for reference}
}
\value{
\item{dataframe}{dataframe for Bradley-Terry run}
\item{ability}{Initial ability vector for iterations}
\item{worstTeam}{The worst team whose ability can be set as 0 during any model's run}
}
\description{
Dataframe initialization
}
\details{
Initial the raw dataframe and return the un-estimated ability vector and the worst
team who losses most
}
