% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NFL2010.R
\docType{data}
\name{NFL2010}
\alias{NFL2010}
\title{The 2010 NFL Regular Season}
\format{A dataframe containing all match results with 5 columns
\describe{
\item{home.team}{Team who plays at home}
\item{away.team}{Team who plays away}
\item{home.win}{Take "1" if home team wins}
\item{away.win}{Take "1" if away team wins}
\item{date}{Number of days until now}
}}
\usage{
NFL2010
}
\description{
A dataframe containing all match results with 5 columns
}
\keyword{datasets}
