% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.swlasso}
\alias{plot.swlasso}
\title{Plot the Lasso path}
\usage{
##S3 method for class "swlasso"
}
\arguments{
\item{x}{Object with class "swlasso"}

\item{...}{Further arguments pass to or from other methods}
}
\description{
Plot the whole lasso path run by BTdecayLasso() with given lambda and path = TRUE
}
