% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.wlasso}
\alias{plot.wlasso}
\title{Plot the Lasso path}
\usage{
##S3 method for class "wlasso"
}
\arguments{
\item{x}{Object with class "wlasso"}

\item{...}{Further arguments pass to or from other methods}
}
\description{
Plot the whole lasso path run by BTdecayLasso() with lambda = NULL and path = TRUE
}
