% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTdecayLasso.R
\name{BTdecayLasso}
\alias{BTdecayLasso}
\title{Bradley-Terry Model with Exponential Decayed weighted likelihood and weighted Lasso with a given Lasso lambda and Lasso weight}
\usage{
BTdecayLasso(dataframe, ability, lambda = NULL, weight = NULL,
  path = TRUE, decay.rate = 0, fixed = 1, thersh = 1e-05, max = 100,
  iter = 100)
}
\arguments{
\item{dataframe}{Matrix with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).
It can be generated using function BTdataframe.}

\item{ability}{A column vector of teams ability, the last row is the home parameter.
The row number is consistent with the team's index shown in dataframe.
It can be generated using function BTdataframe.}

\item{lambda}{The amount of Lasso penalty induced.}

\item{weight}{Weight for Lasso penalty on different abilities.}

\item{path}{whether the whole Lasso path will be run (plot.BTdecayLasso is enabled only if path = TRUE)}

\item{decay.rate}{The exponential decay rate. Usually ranging from (0, 0.1), A larger decay rate weights more
importance to most recent matches and the estimated parameters reflect more on recent behaviour.}

\item{fixed}{A teams index whose ability will be fixed as 0 (usually the team loss most which can be
generated using function BTdataframe).}

\item{thersh}{Thershold for convergency}

\item{max}{Maximum weight for w_{ij} (weight used for Adaptive Lasso)}

\item{iter}{Number of iterations used in L-BFGS-B algorithm.}
}
\value{
\item{ability}{Estimated ability scores with user given lambda}
\item{likelihood}{Negative likelihood of objective function with user given lambda}
\item{df}{Degree of freedom with user given lambda(number of distinct \eqn{\mu})}
\item{penalty}{\eqn{s/max(s)} with user given lambda}
\item{Lambda}{User given lambda}
\item{ability.path}{Estimated ability scores on whole Lasso path}
\item{likelihood.path}{Negative likelihood of objective function on whole Lasso path}
\item{df.path}{Degree of freedom on whole Lasso path(number of distinct \eqn{\mu})}
\item{penalty.path}{\eqn{s/max(s)} on whole Lasso path}
\item{Lambda.path}{Whole Lasso path}
\item{path}{Whether whole Lasso path will be run}
}
\description{
Normal Adaptive Lasso's weight can be determined using funcion BTLasso.weight.
If you want to use equal weight, when there are n teams, you can define a n*n upper triangular matrix with diagonal is 0 and other entries of upper-right part are 1
}
\details{
The objective likelihood function to be optimized is,
\deqn{\sum_{k=1}^{n}\sum_{i<j}\exp(-\alpha t_{k})\cdot(y_{ij}(\tau h_{ij}^{t_{k}}+\mu_{i}-\mu_{j})-\log(1+\exp(\tau h_{ij}^{t_{k}}+\mu_{i}-\mu_{j})))}
With the Lasso constraint,
\deqn{\sum_{i<j}w_{ij}\left|\mu_{i}-\mu_{j}\right|\leq s}
where n is the number of matches, \eqn{\alpha} is the exponential decay rate, \eqn{\tau} is the home parameter and 
\eqn{y_{ij}} takes 0 if i is defeated by j, 1 otherwise. \eqn{\mu_{i}} is the team i's ability score and penalty is 1-s/max(s).
This likelihood function is optimized using L-BFGS-B method with package \bold{optimr}. Without specifying path = FALSE, the whole lasso path will be run
and we can use plot.BTdecayLasso to do a lasso path plot
}
\examples{
##Initializing Dataframe
x <- BTdataframe(NFL2010)

\dontrun{
##BTdecayLasso run with exponential decay rate 0.005 and 
##lambda 0.1 on whole lasso path using adaptive lasso
y1 <- BTdecayLasso(x$dataframe, x$ability, lambda = 0.1, 
                   decay.rate = 0.005, fixed = x$worstTeam)
summary(y1)

##Defining equal weight
n <- nrow(x$ability) - 1
w2 <- matrix(1, nrow = n, ncol = n)
w2[lower.tri(w2, diag = TRUE)] <- 0

##BTdecayLasso run with exponential decay rate 0.005 and with a specific lambda 0.1
y2 <- BTdecayLasso(x$dataframe, x$ability, lambda = 0.1, weight = w2, 
                   path = FALSE, decay.rate = 0.005, fixed = x$worstTeam)
summary(y2)
}

}
\references{
Masarotto, G. and Varin, C.(2012) The Ranking Lasso and its Application to Sport Tournaments. 
*The Annals of Applied Statistics* **6** 1949--1970.

Zou, H. (2006) The adaptive lasso and its oracle properties. 
*J.Amer.Statist.Assoc* **101** 1418--1429.
}
\seealso{
\code{\link{BTdataframe}} for dataframe initialization,
\code{\link{plot}} for Lasso path plot if path = TRUE in this function's run
}
