% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTdecayLassoC.R
\name{BTdecayLassoC}
\alias{BTdecayLassoC}
\title{Bradley-Terry Model with Exponential Decayed weighted likelihood and weighted Lasso with AIC or BIC criteria}
\usage{
BTdecayLassoC(dataframe, ability, weight = NULL, lambda = NULL,
  criteria = "AIC", type = "HYBRID", decay.rate = 0, fixed = 1,
  thersh = 1e-05, iter = 100, max = 100)
}
\arguments{
\item{dataframe}{Matrix with 5 columns. First column is the index of the home teams
(use numbers to denote teams).
Second column is the index of the away teams.
Third column is the number of wins of home teams (usually to be 0/1).
Fourth column is the number of wins of away teams (usually to be 0/1).
Fifth column is the scalar of time when the match is played until now (Time lag).
It can be generated using function BTdataframe.}

\item{ability}{A column vector of teams ability, the last row is the home parameter.
The row number is consistent with the team's index shown in dataframe.
It can be generated using function BTdataframe.}

\item{weight}{Weight for Lasso penalty on different abilities}

\item{criteria}{"AIC" or "BIC"}

\item{type}{"HYBRID" or "LASSO"}

\item{decay.rate}{The exponential decay rate. Usually ranging from (0, 0.1), A larger decay rate weights more
importance to most recent matches and the estimated parameters reflect more on recent behaviour.}

\item{fixed}{A teams index whose ability will be fixed as 0 (usually the team loss most which can be
generated using function BTdataframe).}

\item{thersh}{Thershold for convergency}

\item{iter}{Number of iterations used in L-BFGS-B algorithm.}

\item{max}{Maximum weight for w_{ij} (weight used for Adaptive Lasso)}

\item{Lambda}{A sequence of Lambda}
}
\value{
\item{Score}{Lowest AIC or BIC score}
\item{Optimal.degree}{The degree of freedom where lowest AIC or BIC score is achieved}
\item{Optimal.ability}{The ability where lowest AIC or BIC score is achieved}
\item{ability}{Matrix contains all abilities computed in this algorithm}
\item{Optimal.lambda}{The lambda where lowest score is attained}
\item{Optimal.penalty}{The penalty (s/\eqn{\max(s)}) where lowest score is attained}
\item{type}{Type of model selection method}
}
\description{
Bradley-Terry Model with Exponential Decayed weighted likelihood and weighted Lasso with AIC or BIC criteria
}
\details{
Model selection through AIC or BIC method
}
\examples{
##Initializing Dataframe
x <- BTdataframe(NFL2010)

##Model selection through AIC
w <- BTLasso.weight(NFL$df, NFL$ability)
z <- BTdecayLassoC(NFL$df, NFL$ability, w, criteria = "AIC", type = "LASSO")
}
